Refactor: Enemy.cs is not base class, move content to EnemyHammerBot
Move previous Enemy.cs to EnemyHammerBot.cs and refactor Enemy.cs to be the base class for all enemies.
